ACTUAL FIX

+++ Microwave.swift
@@ -59,10 +59,10 @@ if FOOD_CHOICES.contains(userFoodChoice) {
                 }
                 // Apply the quantity formula on the base reheat time.
                 // To get the actual reheat time. [in seconds]
-                let actualReheatTime = baseReheatTime * (quantity + 1) * 0.5
+                let actualReheatTime = Double(baseReheatTime) * Double(quantity + 1) * 0.5
                 // Convert pure seconds to a combination of minutes and seconds.
-                let minutes = actualReheatTime / SECONDS_PER_MINUTE
-                let seconds = actualReheatTime % SECONDS_PER_MINUTE
+                let minutes = Int(actualReheatTime / Double(SECONDS_PER_MINUTE))
+                let seconds = Int(actualReheatTime) % SECONDS_PER_MINUTE
                 // Display the result. [IN BLUE]
                 print("\u{001B}[0;34mReheat time: \(minutes)", terminator: "")
                 print(" minute(s) and \(seconds) second(s).")
